Change quantity to uint64

+++ models/subscription.go
@@ -17,7 +17,7 @@ type Subscription struct {
 	Frequency string             `json:"frequency"`
 	RoasterID uuid.UUID          `json:"roasterId"`
 	ItemID    uuid.UUID          `json:"itemId"`
-	Quantity  int                `json:"quantity"`
+	Quantity  uint64                `json:"quantity"`
 }
 
 /*RequestSubscription represents the data needed to create a new subscription entry*/
@@ -26,17 +26,17 @@ type RequestSubscription struct {
 	Frequency string    `json:"frequency" binding:"required"`
 	RoasterID uuid.UUID `json:"roasterId" binding:"required"`
 	ItemID    uuid.UUID `json:"itemId" binding:"required"`
-	Quantity  int       `json:"quantity" binding:"required"`
+	Quantity  uint64       `json:"quantity" binding:"required"`
 }
 
-/*RequestOrder represents the data sent from Coinage to create a new order entry wihin Warehouse*/
+/*RequestOrder represents the data needed to create a new order entry wihin Warehouse*/
 type RequestOrder struct {
 	UserID uuid.UUID `json:"userId binding:"required"`
 	ItemID uuid.UUID `json:"itemId binding:"required"`
 }
 
 /*NewSubscription creates a new subscription with a new uuid*/
-func NewSubscription(userID uuid.UUID, frequency string, roasterID uuid.UUID, itemID uuid.UUID, quantity int) *Subscription {
+func NewSubscription(userID uuid.UUID, frequency string, roasterID uuid.UUID, itemID uuid.UUID, quantity uint64) *Subscription {
 	return &Subscription{
 		ID:        uuid.NewUUID(),
 		UserID:    userID,
